Move product related fixtures from global fixtures file
Move fixtures from other files
