+++ saleor/graphql/product/tests/queries/test_products_query_with_filter.py
@@ -27,31 +27,27 @@ from .....tests.utils import dummy_editorjs
 from .....warehouse.models import Allocation, Reservation, Stock, Warehouse
 from ....tests.utils import get_graphql_content
 
-
-@pytest.fixture
-def query_products_with_filter():
-    query = """
-        query ($filter: ProductFilterInput!, $channel: String) {
-          products(first:5, filter: $filter, channel: $channel) {
-            edges{
-              node{
-                id
-                name
-              }
-            }
+QUERY_PRODUCTS_WITH_FILTER = """
+    query ($filter: ProductFilterInput!, $channel: String) {
+      products(first:5, filter: $filter, channel: $channel) {
+        edges{
+          node{
+            id
+            name
           }
         }
-        """
-    return query
+      }
+    }
+    """
 
 
 def test_products_query_with_filter_attributes(
-    query_products_with_filter,
     staff_api_client,
     product,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     product_type = ProductType.objects.create(
         name="Custom Type",
         slug="custom-type",
@@ -81,8 +77,12 @@ def test_products_query_with_filter_attributes(
     }
 
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -109,13 +109,13 @@ def test_products_query_with_filter_numeric_attributes(
     gte,
     lte,
     expected_products_index,
-    query_products_with_filter,
     staff_api_client,
     product,
     category,
     numeric_attribute,
     permission_manage_products,
 ):
+    # given
     product.product_type.product_attributes.add(numeric_attribute)
     associate_attribute_values_to_instance(
         product,
@@ -181,8 +181,12 @@ def test_products_query_with_filter_numeric_attributes(
     }
 
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == len(expected_products_index)
@@ -204,13 +208,13 @@ def test_products_query_with_filter_numeric_attributes(
 def test_products_query_with_filter_boolean_attributes(
     filter_value,
     expected_products_index,
-    query_products_with_filter,
     staff_api_client,
     product,
     category,
     boolean_attribute,
     permission_manage_products,
 ):
+    # given
     product.product_type.product_attributes.add(boolean_attribute)
 
     associate_attribute_values_to_instance(
@@ -251,8 +255,12 @@ def test_products_query_with_filter_boolean_attributes(
     }
 
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == len(expected_products_index)
@@ -273,13 +281,11 @@ def test_products_query_with_filter_boolean_attributes(
 )
 def test_products_query_with_filter_non_existing_boolean_attributes(
     filter_value,
-    query_products_with_filter,
     staff_api_client,
     product,
     permission_manage_products,
 ):
     # given
-
     variables = {
         "filter": {
             "attributes": [{"slug": "non-existing-atr", "boolean": filter_value}]
@@ -289,7 +295,7 @@ def test_products_query_with_filter_non_existing_boolean_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -306,7 +312,6 @@ def test_products_query_with_filter_non_existing_boolean_attributes(
 )
 def test_products_query_with_filter_boolean_attributes_not_assigned_to_product(
     filter_value,
-    query_products_with_filter,
     staff_api_client,
     product,
     category,
@@ -325,7 +330,7 @@ def test_products_query_with_filter_boolean_attributes_not_assigned_to_product(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -334,13 +339,13 @@ def test_products_query_with_filter_boolean_attributes_not_assigned_to_product(
 
 
 def test_products_query_with_filter_by_attributes_values_and_range(
-    query_products_with_filter,
     staff_api_client,
     product,
     category,
     numeric_attribute,
     permission_manage_products,
 ):
+    # given
     product_attr = get_product_attributes(product).first()
     attr_value_1 = get_product_attribute_values(product, product_attr).first()
     product.product_type.product_attributes.add(numeric_attribute)
@@ -385,8 +390,12 @@ def test_products_query_with_filter_by_attributes_values_and_range(
     }
 
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -397,13 +406,13 @@ def test_products_query_with_filter_by_attributes_values_and_range(
 
 
 def test_products_query_with_filter_swatch_attributes(
-    query_products_with_filter,
     staff_api_client,
     product,
     category,
     swatch_attribute,
     permission_manage_products,
 ):
+    # given
     product.product_type.product_attributes.add(swatch_attribute)
     associate_attribute_values_to_instance(
         product,
@@ -444,8 +453,12 @@ def test_products_query_with_filter_swatch_attributes(
     }
 
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -455,7 +468,6 @@ def test_products_query_with_filter_swatch_attributes(
 
 
 def test_products_query_with_filter_date_range_date_attributes(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -506,7 +518,7 @@ def test_products_query_with_filter_date_range_date_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -519,7 +531,6 @@ def test_products_query_with_filter_date_range_date_attributes(
 
 
 def test_products_query_with_filter_date_range_date_variant_attributes(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -570,7 +581,7 @@ def test_products_query_with_filter_date_range_date_variant_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -583,7 +594,6 @@ def test_products_query_with_filter_date_range_date_variant_attributes(
 
 
 def test_products_query_with_filter_date_range_date_time_attributes(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -637,7 +647,7 @@ def test_products_query_with_filter_date_range_date_time_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -650,7 +660,6 @@ def test_products_query_with_filter_date_range_date_time_attributes(
 
 
 def test_products_query_with_filter_date_range_date_time_variant_attributes(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -704,7 +713,7 @@ def test_products_query_with_filter_date_range_date_time_variant_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -717,7 +726,6 @@ def test_products_query_with_filter_date_range_date_time_variant_attributes(
 
 
 def test_products_query_with_filter_date_time_range_date_time_attributes(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -778,7 +786,7 @@ def test_products_query_with_filter_date_time_range_date_time_attributes(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -791,21 +799,22 @@ def test_products_query_with_filter_date_time_range_date_time_attributes(
 
 
 def test_products_query_filter_by_non_existing_attribute(
-    query_products_with_filter, api_client, product_list, channel_USD
+    api_client, product_list, channel_USD
 ):
     variables = {
         "channel": channel_USD.slug,
         "filter": {"attributes": [{"slug": "i-do-not-exist", "values": ["red"]}]},
     }
-    response = api_client.post_graphql(query_products_with_filter, variables)
+    response = api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
     products = content["data"]["products"]["edges"]
     assert len(products) == 0
 
 
 def test_products_query_with_filter_category(
-    query_products_with_filter, staff_api_client, product, permission_manage_products
+    staff_api_client, product, permission_manage_products
 ):
+    # given
     category = Category.objects.create(name="Custom", slug="custom")
     second_product = product
     second_product.id = None
@@ -815,8 +824,12 @@ def test_products_query_with_filter_category(
 
     category_id = graphene.Node.to_global_id("Category", category.id)
     variables = {"filter": {"categories": [category_id]}}
+
+    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
+
+    # then
     content = get_graphql_content(response)
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
@@ -827,8 +840,9 @@ def test_products_query_with_filter_category(
 
 
 def test_products_query_with_filter_has_category_false(
-    query_products_with_filter, staff_api_client, product, permission_manage_products
+    staff_api_client, product, permission_manage_products
 ):
+    # given
     second_product = product
     second_product.category = None
     second_product.id = None
@@ -837,8 +851,12 @@ def test_products_query_with_filter_has_category_false(
 
     variables = {"filter": {"hasCategory": False}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -848,11 +866,11 @@ def test_products_query_with_filter_has_category_false(
 
 
 def test_products_query_with_filter_has_category_true(
-    query_products_with_filter,
     staff_api_client,
     product_without_category,
     permission_manage_products,
 ):
+    # given
     category = Category.objects.create(name="Custom", slug="custom")
     second_product = product_without_category
     second_product.category = category
@@ -862,8 +880,12 @@ def test_products_query_with_filter_has_category_true(
 
     variables = {"filter": {"hasCategory": True}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -873,12 +895,12 @@ def test_products_query_with_filter_has_category_true(
 
 
 def test_products_query_with_filter_collection(
-    query_products_with_filter,
     staff_api_client,
     product,
     collection,
     permission_manage_products,
 ):
+    # given
     second_product = product
     second_product.id = None
     second_product.slug = "second-product"
@@ -888,8 +910,12 @@ def test_products_query_with_filter_collection(
     collection_id = graphene.Node.to_global_id("Collection", collection.id)
     variables = {"filter": {"collections": [collection_id]}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -899,11 +925,11 @@ def test_products_query_with_filter_collection(
 
 
 def test_products_query_with_filter_category_and_search(
-    query_products_with_filter,
     staff_api_client,
     product,
     permission_manage_products,
 ):
+    # given
     category = Category.objects.create(name="Custom", slug="custom")
     second_product = product
     second_product.id = None
@@ -922,8 +948,12 @@ def test_products_query_with_filter_category_and_search(
     category_id = graphene.Node.to_global_id("Category", category.id)
     variables = {"filter": {"categories": [category_id], "search": product.name}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product.id)
     products = content["data"]["products"]["edges"]
 
@@ -933,7 +963,6 @@ def test_products_query_with_filter_category_and_search(
 
 
 def test_products_query_with_filter_gift_card_false(
-    query_products_with_filter,
     staff_api_client,
     product,
     shippable_gift_card_product,
@@ -944,7 +973,7 @@ def test_products_query_with_filter_gift_card_false(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -957,7 +986,6 @@ def test_products_query_with_filter_gift_card_false(
 
 
 def test_products_query_with_filter_gift_card_true(
-    query_products_with_filter,
     staff_api_client,
     product,
     shippable_gift_card_product,
@@ -968,7 +996,7 @@ def test_products_query_with_filter_gift_card_true(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -990,12 +1018,12 @@ def test_products_query_with_filter_gift_card_true(
 )
 def test_products_query_with_filter(
     products_filter,
-    query_products_with_filter,
     staff_api_client,
     product,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     assert "Juice1" not in product.name
 
     second_product = product
@@ -1026,8 +1054,12 @@ def test_products_query_with_filter(
     second_product.save(update_fields=["search_vector"])
     variables = {"filter": products_filter, "channel": channel_USD.slug}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     second_product_id = graphene.Node.to_global_id("Product", second_product.id)
     products = content["data"]["products"]["edges"]
 
@@ -1037,12 +1069,12 @@ def test_products_query_with_filter(
 
 
 def test_products_query_with_price_filter_as_staff(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     product = product_list[0]
     product.variants.first().channel_listings.filter().update(price_amount=None)
 
@@ -1051,20 +1083,24 @@ def test_products_query_with_price_filter_as_staff(
         "channel": channel_USD.slug,
     }
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 3
 
 
 def test_products_query_with_price_filter_as_user(
-    query_products_with_filter,
     user_api_client,
     product_list,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     product = product_list[0]
     product.variants.first().channel_listings.filter().update(price_amount=None)
     second_product_id = graphene.Node.to_global_id("Product", product_list[1].id)
@@ -1073,8 +1109,12 @@ def test_products_query_with_price_filter_as_user(
         "filter": {"price": {"gte": 9, "lte": 31}},
         "channel": channel_USD.slug,
     }
-    response = user_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = user_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 2
@@ -1085,20 +1125,24 @@ def test_products_query_with_price_filter_as_user(
 @pytest.mark.parametrize("is_published", [(True), (False)])
 def test_products_query_with_filter_search_by_sku(
     is_published,
-    query_products_with_filter,
     staff_api_client,
     product_with_two_variants,
     product_with_default_variant,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     ProductChannelListing.objects.filter(
         product=product_with_default_variant, channel=channel_USD
     ).update(is_published=is_published)
     variables = {"filter": {"search": "1234"}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product_with_default_variant.id)
     products = content["data"]["products"]["edges"]
 
@@ -1110,7 +1154,6 @@ def test_products_query_with_filter_search_by_sku(
 @pytest.mark.parametrize("search_value", ["new", "NEW color", "Color"])
 def test_products_query_with_filter_search_by_dropdown_attribute_value(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1141,12 +1184,13 @@ def test_products_query_with_filter_search_by_dropdown_attribute_value(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1162,7 +1206,6 @@ def test_products_query_with_filter_search_by_dropdown_attribute_value(
 )
 def test_products_query_with_filter_search_by_multiselect_attribute_value(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1201,13 +1244,13 @@ def test_products_query_with_filter_search_by_multiselect_attribute_value(
     product_with_multiselect_attr.save(update_fields=["search_vector"])
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
-
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1220,7 +1263,6 @@ def test_products_query_with_filter_search_by_multiselect_attribute_value(
 @pytest.mark.parametrize("search_value", ["rich", "test rich", "RICH text"])
 def test_products_query_with_filter_search_by_rich_text_attribute(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1251,12 +1293,13 @@ def test_products_query_with_filter_search_by_rich_text_attribute(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1269,7 +1312,6 @@ def test_products_query_with_filter_search_by_rich_text_attribute(
 @pytest.mark.parametrize("search_value", ["plain", "test plain", "PLAIN text"])
 def test_products_query_with_filter_search_by_plain_text_attribute(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1300,12 +1342,13 @@ def test_products_query_with_filter_search_by_plain_text_attribute(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1318,7 +1361,6 @@ def test_products_query_with_filter_search_by_plain_text_attribute(
 @pytest.mark.parametrize("search_value", ["13456", "13456 cm"])
 def test_products_query_with_filter_search_by_numeric_attribute_value(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1352,12 +1394,13 @@ def test_products_query_with_filter_search_by_numeric_attribute_value(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1368,7 +1411,6 @@ def test_products_query_with_filter_search_by_numeric_attribute_value(
 
 
 def test_products_query_with_filter_search_by_numeric_attribute_value_without_unit(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1400,12 +1442,13 @@ def test_products_query_with_filter_search_by_numeric_attribute_value_without_un
 
     variables = {"filter": {"search": "13456"}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1418,7 +1461,6 @@ def test_products_query_with_filter_search_by_numeric_attribute_value_without_un
 @pytest.mark.parametrize("search_value", ["2020", "2020-10-10"])
 def test_products_query_with_filter_search_by_date_attribute_value(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1449,12 +1491,13 @@ def test_products_query_with_filter_search_by_date_attribute_value(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1467,7 +1510,6 @@ def test_products_query_with_filter_search_by_date_attribute_value(
 @pytest.mark.parametrize("search_value", ["2020", "2020-10-10", "22:20"])
 def test_products_query_with_filter_search_by_date_time_attribute_value(
     search_value,
-    query_products_with_filter,
     staff_api_client,
     product_list,
     permission_manage_products,
@@ -1500,12 +1542,13 @@ def test_products_query_with_filter_search_by_date_time_attribute_value(
 
     variables = {"filter": {"search": search_value}, "channel": channel_USD.slug}
 
-    # when
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
-    # then
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
@@ -1516,59 +1559,67 @@ def test_products_query_with_filter_search_by_date_time_attribute_value(
 
 
 def test_products_query_with_is_published_filter_variants_without_prices(
-    query_products_with_filter,
     staff_api_client,
     variant,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     ProductVariantChannelListing.objects.filter(
         variant__product=variant.product
     ).update(price_amount=None)
 
     variables = {"channel": channel_USD.slug, "filter": {"isPublished": True}}
+
+    # when
     response = staff_api_client.post_graphql(
-        query_products_with_filter,
+        QUERY_PRODUCTS_WITH_FILTER,
         variables,
         permissions=[permission_manage_products],
         check_no_permissions=False,
     )
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 0
 
 
 def test_products_query_with_is_published_filter_one_variant_without_price(
-    query_products_with_filter,
     staff_api_client,
     variant,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     variant.channel_listings.update(price_amount=None)
 
     variables = {"channel": channel_USD.slug, "filter": {"isPublished": True}}
+
+    # when
     response = staff_api_client.post_graphql(
-        query_products_with_filter,
+        QUERY_PRODUCTS_WITH_FILTER,
         variables,
         permissions=[permission_manage_products],
         check_no_permissions=False,
     )
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 1
 
 
 def test_products_query_with_filter_stock_availability_as_staff(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     order_line,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     for product in product_list:
         stock = product.variants.first().stocks.first()
         Allocation.objects.create(
@@ -1583,15 +1634,18 @@ def test_products_query_with_filter_stock_availability_as_staff(
         "channel": channel_USD.slug,
     }
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 3
 
 
 def test_products_query_with_filter_stock_availability_including_reservations(
-    query_products_with_filter,
     staff_api_client,
     product_list,
     order_line,
@@ -1640,7 +1694,7 @@ def test_products_query_with_filter_stock_availability_including_reservations(
     staff_api_client.user.user_permissions.add(permission_manage_products)
 
     # when
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -1652,13 +1706,13 @@ def test_products_query_with_filter_stock_availability_including_reservations(
 
 
 def test_products_query_with_filter_stock_availability_as_user(
-    query_products_with_filter,
     user_api_client,
     product_list,
     order_line,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     for product in product_list:
         stock = product.variants.first().stocks.first()
         Allocation.objects.create(
@@ -1672,8 +1726,12 @@ def test_products_query_with_filter_stock_availability_as_user(
         "filter": {"stockAvailability": "OUT_OF_STOCK"},
         "channel": channel_USD.slug,
     }
-    response = user_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = user_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product_list[1].id)
     second_product_id = graphene.Node.to_global_id("Product", product_list[2].id)
 
@@ -1687,20 +1745,24 @@ def test_products_query_with_filter_stock_availability_as_user(
 
 
 def test_products_query_with_filter_stock_availability_channel_without_shipping_zones(
-    query_products_with_filter,
     staff_api_client,
     product,
     permission_manage_products,
     channel_USD,
 ):
+    # given
     channel_USD.shipping_zones.clear()
     variables = {
         "filter": {"stockAvailability": "OUT_OF_STOCK"},
         "channel": channel_USD.slug,
     }
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
     product_id = graphene.Node.to_global_id("Product", product.id)
 
@@ -1709,7 +1771,6 @@ def test_products_query_with_filter_stock_availability_channel_without_shipping_
 
 
 def test_products_query_with_filter_stock_availability_only_stock_in_cc_warehouse(
-    query_products_with_filter,
     user_api_client,
     product,
     order_line,
@@ -1730,7 +1791,7 @@ def test_products_query_with_filter_stock_availability_only_stock_in_cc_warehous
     }
 
     # when
-    response = user_api_client.post_graphql(query_products_with_filter, variables)
+    response = user_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -1762,13 +1823,13 @@ def test_products_query_with_filter_stocks(
     warehouse_indexes,
     count,
     indexes_of_products_in_result,
-    query_products_with_filter,
     staff_api_client,
     product_with_single_variant,
     product_with_two_variants,
     warehouse,
     channel_USD,
 ):
+    # given
     product1 = product_with_single_variant
     product2 = product_with_two_variants
     products = [product1, product2]
@@ -1820,10 +1881,14 @@ def test_products_query_with_filter_stocks(
         },
         "channel": channel_USD.slug,
     }
+
+    # when
     response = staff_api_client.post_graphql(
-        query_products_with_filter, variables, check_no_permissions=False
+        QUERY_PRODUCTS_WITH_FILTER, variables, check_no_permissions=False
     )
     content = get_graphql_content(response)
+
+    # then
     products_data = content["data"]["products"]["edges"]
 
     product_ids = {
@@ -1835,9 +1900,7 @@ def test_products_query_with_filter_stocks(
     assert {node["node"]["id"] for node in products_data} == product_ids
 
 
-def test_query_products_with_filter_ids(
-    api_client, product_list, query_products_with_filter, channel_USD
-):
+def test_query_products_with_filter_ids(api_client, product_list, channel_USD):
     # given
     product_ids = [
         graphene.Node.to_global_id("Product", product.id) for product in product_list
@@ -1848,7 +1911,7 @@ def test_query_products_with_filter_ids(
     }
 
     # when
-    response = api_client.post_graphql(query_products_with_filter, variables)
+    response = api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
 
     # then
     content = get_graphql_content(response)
@@ -1859,17 +1922,21 @@ def test_query_products_with_filter_ids(
 
 
 def test_products_query_with_filter_has_preordered_variants_false(
-    query_products_with_filter,
     staff_api_client,
     preorder_variant_global_threshold,
     product_without_shipping,
     permission_manage_products,
 ):
+    # given
     product = product_without_shipping
     variables = {"filter": {"hasPreorderedVariants": False}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product.id)
     products = content["data"]["products"]["edges"]
 
@@ -1879,17 +1946,21 @@ def test_products_query_with_filter_has_preordered_variants_false(
 
 
 def test_products_query_with_filter_has_preordered_variants_true(
-    query_products_with_filter,
     staff_api_client,
     preorder_variant_global_threshold,
     product_without_shipping,
     permission_manage_products,
 ):
+    # given
     product = preorder_variant_global_threshold.product
     variables = {"filter": {"hasPreorderedVariants": True}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product.id)
     products = content["data"]["products"]["edges"]
 
@@ -1899,11 +1970,11 @@ def test_products_query_with_filter_has_preordered_variants_true(
 
 
 def test_products_query_with_filter_has_preordered_variants_before_end_date(
-    query_products_with_filter,
     staff_api_client,
     preorder_variant_global_threshold,
     permission_manage_products,
 ):
+    # given
     variant = preorder_variant_global_threshold
     variant.preorder_end_date = timezone.now() + datetime.timedelta(days=3)
     variant.save(update_fields=["preorder_end_date"])
@@ -1911,8 +1982,12 @@ def test_products_query_with_filter_has_preordered_variants_before_end_date(
     product = preorder_variant_global_threshold.product
     variables = {"filter": {"hasPreorderedVariants": True}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     product_id = graphene.Node.to_global_id("Product", product.id)
     products = content["data"]["products"]["edges"]
 
@@ -1922,19 +1997,23 @@ def test_products_query_with_filter_has_preordered_variants_before_end_date(
 
 
 def test_products_query_with_filter_has_preordered_variants_after_end_date(
-    query_products_with_filter,
     staff_api_client,
     preorder_variant_global_threshold,
     permission_manage_products,
 ):
+    # given
     variant = preorder_variant_global_threshold
     variant.preorder_end_date = timezone.now() - datetime.timedelta(days=3)
     variant.save(update_fields=["preorder_end_date"])
 
     variables = {"filter": {"hasPreorderedVariants": True}}
     staff_api_client.user.user_permissions.add(permission_manage_products)
-    response = staff_api_client.post_graphql(query_products_with_filter, variables)
+
+    # when
+    response = staff_api_client.post_graphql(QUERY_PRODUCTS_WITH_FILTER, variables)
     content = get_graphql_content(response)
+
+    # then
     products = content["data"]["products"]["edges"]
 
     assert len(products) == 0
