+++ saleor/product/tests/fixtures/__init__.py
@@ -1,0 +1,7 @@
+from .category import *  # noqa: F403
+from .collection import *  # noqa: F403
+from .digital_content import *  # noqa: F403
+from .product import *  # noqa: F403
+from .product_media import *  # noqa: F403
+from .product_type import *  # noqa: F403
+from .variant import *  # noqa: F403
