+++ saleor/tests/fixtures.py
@@ -1,6 +1,4 @@
 import datetime
-import itertools
-import random
 import uuid
 from contextlib import contextmanager
 from decimal import Decimal
@@ -13,7 +11,6 @@ import graphene
 import pytest
 from django.conf import settings
 from django.contrib.sites.models import Site
-from django.core.files import File
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.db import connection
 from django.template.defaultfilters import truncatechars
@@ -43,7 +40,6 @@ from ..checkout.utils import (
 from ..core import JobStatus
 from ..core.models import EventDelivery, EventDeliveryAttempt, EventPayload
 from ..core.payments import PaymentInterface
-from ..core.postgres import FlatConcatSearchVector
 from ..core.units import MeasurementUnits
 from ..core.utils.editorjs import clean_editor_js
 from ..csv.events import ExportEvents
@@ -83,27 +79,13 @@ from ..payment.utils import create_manual_adjustment_events
 from ..permission.enums import get_permissions
 from ..permission.models import Permission
 from ..plugins.manager import get_plugins_manager
-from ..product import ProductMediaTypes, ProductTypeKind
 from ..product.models import (
-    Category,
     CategoryTranslation,
-    Collection,
-    CollectionChannelListing,
     CollectionTranslation,
-    DigitalContent,
-    DigitalContentUrl,
-    Product,
-    ProductChannelListing,
-    ProductMedia,
     ProductTranslation,
-    ProductType,
-    ProductVariant,
     ProductVariantChannelListing,
     ProductVariantTranslation,
-    VariantMedia,
 )
-from ..product.search import prepare_product_search_vector_value
-from ..product.tests.utils import create_image
 from ..product.utils.variants import fetch_variants_for_promotion_rules
 from ..shipping.models import (
     ShippingMethod,
@@ -118,7 +100,6 @@ from ..tax import TaxCalculationStrategy
 from ..warehouse.models import (
     PreorderReservation,
     Reservation,
-    Stock,
     Warehouse,
 )
 from ..webhook.event_types import WebhookEventAsyncType
@@ -1190,74 +1171,6 @@ def user_checkouts(request, user_checkout_with_items, user_checkout_with_items_f
 
 
 @pytest.fixture
-def product_type_generator(
-    attribute_generator, attribute_value_generator, default_tax_class
-):
-    def create_product_type(
-        name="Default Type",
-        slug="default-type",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-        tax_class=default_tax_class,
-        product_attributes=None,
-        variant_attributes=None,
-    ):
-        product_type = ProductType.objects.create(
-            name=name,
-            slug=slug,
-            kind=kind,
-            has_variants=has_variants,
-            is_shipping_required=is_shipping_required,
-            tax_class=tax_class,
-        )
-        if product_attributes is None:
-            product_attribute = attribute_generator(
-                external_reference="colorAttributeExternalReference",
-                slug="color",
-                name="Color",
-                type=AttributeType.PRODUCT_TYPE,
-                filterable_in_storefront=True,
-                filterable_in_dashboard=True,
-                available_in_grid=True,
-            )
-
-            attribute_value_generator(
-                external_reference="colorAttributeValue1ExternalReference",
-                name="Red",
-                slug="red",
-                attribute=product_attribute,
-            )
-
-            product_attributes = [product_attribute]
-        if variant_attributes is None:
-            variant_attribute = attribute_generator(
-                external_reference="sizeAttributeExternalReference",
-                slug="size",
-                name="Size",
-                type=AttributeType.PRODUCT_TYPE,
-                filterable_in_storefront=True,
-                filterable_in_dashboard=True,
-                available_in_grid=True,
-            )
-
-            attribute_value_generator(
-                name="Small",
-                slug="small",
-                attribute=variant_attribute,
-            )
-            variant_attributes = [variant_attribute]
-
-        product_type.product_attributes.add(*product_attributes)
-        product_type.variant_attributes.add(
-            *variant_attributes, through_defaults={"variant_selection": True}
-        )
-        return product_type
-
-    return create_product_type
-
-
-@pytest.fixture
 def admin_user(db):
     """Return a Django admin user."""
     return User.objects.create_user(
@@ -1730,56 +1643,6 @@ def multiselect_attribute(db, attribute_generator, attribute_values_generator):
     )
 
     return attribute
-
-
-@pytest.fixture
-def product_type_with_product_attributes(attribute_without_values):
-    product_type = ProductType.objects.create(
-        name="product_type_with_product_attributes",
-        slug="product-type-with-product-attributes",
-        has_variants=False,
-        is_shipping_required=False,
-        weight=0,
-    )
-    product_type.product_attributes.add(attribute_without_values)
-    return product_type
-
-
-@pytest.fixture
-def product_type_with_variant_attributes(attribute_without_values):
-    product_type = ProductType.objects.create(
-        name="product_type_with_variant_attributes",
-        slug="product-type-with-variant-attributes",
-        has_variants=False,
-        is_shipping_required=False,
-        weight=0,
-    )
-    product_type.variant_attributes.add(attribute_without_values)
-    return product_type
-
-
-@pytest.fixture
-def product_with_product_attributes(
-    product_type_with_product_attributes, non_default_category
-):
-    return Product.objects.create(
-        name="product_with_product_attributes",
-        slug="product-with-product-attributes",
-        product_type=product_type_with_product_attributes,
-        category=non_default_category,
-    )
-
-
-@pytest.fixture
-def product_with_variant_attributes(
-    product_type_with_variant_attributes, non_default_category
-):
-    return Product.objects.create(
-        name="product_with_variant_attributes",
-        slug="product-with-variant-attributes",
-        product_type=product_type_with_variant_attributes,
-        category=non_default_category,
-    )
 
 
 @pytest.fixture
@@ -2343,116 +2206,6 @@ def image_list():
 
 
 @pytest.fixture
-def category_generator():
-    def create_category(
-        name="Default",
-        slug="default",
-    ):
-        category = Category.objects.create(
-            name=name,
-            slug=slug,
-        )
-        return category
-
-    return create_category
-
-
-@pytest.fixture
-def category(category_generator):  # pylint: disable=W0613
-    return category_generator()
-
-
-@pytest.fixture
-def category_with_image(db, image, media_root):  # pylint: disable=W0613
-    return Category.objects.create(
-        name="Default2", slug="default2", background_image=image
-    )
-
-
-@pytest.fixture
-def categories(db):
-    category1 = Category.objects.create(name="Category1", slug="cat1")
-    category2 = Category.objects.create(name="Category2", slug="cat2")
-    return [category1, category2]
-
-
-@pytest.fixture
-def category_list():
-    category_1 = Category.objects.create(name="Category 1", slug="category-1")
-    category_2 = Category.objects.create(name="Category 2", slug="category-2")
-    category_3 = Category.objects.create(name="Category 3", slug="category-3")
-    return category_1, category_2, category_3
-
-
-@pytest.fixture
-def categories_tree(db, product_type, channel_USD):  # pylint: disable=W0613
-    parent = Category.objects.create(name="Parent", slug="parent")
-    parent.children.create(name="Child", slug="child")
-    child = parent.children.first()
-
-    product_attr = product_type.product_attributes.first()
-    attr_value = product_attr.values.first()
-
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-10",
-        product_type=product_type,
-        category=child,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-    )
-
-    associate_attribute_values_to_instance(product, {product_attr.pk: [attr_value]})
-    return parent
-
-
-@pytest.fixture
-def categories_tree_with_published_products(
-    categories_tree, product, channel_USD, channel_PLN
-):
-    parent = categories_tree
-    parent_product = product
-    parent_product.category = parent
-
-    child = parent.children.first()
-    child_product = child.products.first()
-
-    product_list = [child_product, parent_product]
-
-    ProductChannelListing.objects.filter(product__in=product_list).delete()
-    product_channel_listings = []
-    for product in product_list:
-        product.save()
-        product_channel_listings.append(
-            ProductChannelListing(
-                product=product,
-                channel=channel_USD,
-                published_at=datetime.datetime.now(tz=datetime.UTC),
-                is_published=True,
-            )
-        )
-        product_channel_listings.append(
-            ProductChannelListing(
-                product=product,
-                channel=channel_PLN,
-                published_at=datetime.datetime.now(tz=datetime.UTC),
-                is_published=True,
-            )
-        )
-    ProductChannelListing.objects.bulk_create(product_channel_listings)
-    return parent
-
-
-@pytest.fixture
-def non_default_category(db):  # pylint: disable=W0613
-    return Category.objects.create(name="Not default", slug="not-default")
-
-
-@pytest.fixture
 def permission_manage_discounts():
     return Permission.objects.get(codename="manage_discounts")
 
@@ -2505,1428 +2258,6 @@ def permission_manage_observability():
 @pytest.fixture
 def permission_manage_taxes():
     return Permission.objects.get(codename="manage_taxes")
-
-
-@pytest.fixture
-def product_type(product_type_generator):
-    return product_type_generator()
-
-
-@pytest.fixture
-def product_type_with_value_required_attributes(
-    color_attribute, size_attribute, default_tax_class
-):
-    product_type = ProductType.objects.create(
-        name="Default Type",
-        slug="default-type",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-        tax_class=default_tax_class,
-    )
-    color_attribute.value_required = True
-    size_attribute.value_required = True
-    Attribute.objects.bulk_update([color_attribute, size_attribute], ["value_required"])
-    product_type.product_attributes.add(color_attribute)
-    product_type.product_attributes.add(size_attribute)
-    return product_type
-
-
-@pytest.fixture
-def product_type_list():
-    product_type_1 = ProductType.objects.create(
-        name="Type 1", slug="type-1", kind=ProductTypeKind.NORMAL
-    )
-    product_type_2 = ProductType.objects.create(
-        name="Type 2", slug="type-2", kind=ProductTypeKind.NORMAL
-    )
-    product_type_3 = ProductType.objects.create(
-        name="Type 3", slug="type-3", kind=ProductTypeKind.NORMAL
-    )
-    return product_type_1, product_type_2, product_type_3
-
-
-@pytest.fixture
-def non_shippable_gift_card_product_type(db):
-    product_type = ProductType.objects.create(
-        name="Gift card type no shipping",
-        slug="gift-card-type-no-shipping",
-        kind=ProductTypeKind.GIFT_CARD,
-        has_variants=True,
-        is_shipping_required=False,
-    )
-    return product_type
-
-
-@pytest.fixture
-def shippable_gift_card_product_type(db):
-    product_type = ProductType.objects.create(
-        name="Gift card type with shipping",
-        slug="gift-card-type-with-shipping",
-        kind=ProductTypeKind.GIFT_CARD,
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    return product_type
-
-
-@pytest.fixture
-def product_type_with_rich_text_attribute(rich_text_attribute):
-    product_type = ProductType.objects.create(
-        name="Default Type",
-        slug="default-type",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.product_attributes.add(rich_text_attribute)
-    product_type.variant_attributes.add(rich_text_attribute)
-    return product_type
-
-
-@pytest.fixture
-def product_type_without_variant():
-    product_type = ProductType.objects.create(
-        name="Type",
-        slug="type",
-        has_variants=False,
-        is_shipping_required=True,
-        kind=ProductTypeKind.NORMAL,
-    )
-    return product_type
-
-
-@pytest.fixture
-def product(product_type, category, warehouse, channel_USD, default_tax_class):
-    product_attr = product_type.product_attributes.first()
-    product_attr_value = product_attr.values.first()
-
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-11",
-        product_type=product_type,
-        category=category,
-        tax_class=default_tax_class,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        discounted_price_amount="10.00",
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    associate_attribute_values_to_instance(
-        product, {product_attr.pk: [product_attr_value]}
-    )
-
-    variant_attr = product_type.variant_attributes.first()
-    variant_attr_value = variant_attr.values.first()
-
-    variant = ProductVariant.objects.create(product=product, sku="123")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=10)
-
-    associate_attribute_values_to_instance(
-        variant, {variant_attr.pk: [variant_attr_value]}
-    )
-
-    return product
-
-
-@pytest.fixture
-def product_with_translations(product):
-    product.translations.create(language_code="pl", name="OldProduct PL")
-    product.translations.create(language_code="de", name="OldProduct DE")
-
-    return product
-
-
-@pytest.fixture
-def shippable_gift_card_product(
-    shippable_gift_card_product_type, category, warehouse, channel_USD
-):
-    product_type = shippable_gift_card_product_type
-
-    product = Product.objects.create(
-        name="Shippable gift card",
-        slug="shippable-gift-card",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        discounted_price_amount="100.00",
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variant = ProductVariant.objects.create(
-        product=product, sku="958", track_inventory=False
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(100),
-        discounted_price_amount=Decimal(100),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=1)
-
-    return product
-
-
-@pytest.fixture
-def product_price_0(category, warehouse, channel_USD):
-    product_type = ProductType.objects.create(
-        name="Type with no shipping",
-        slug="no-shipping",
-        has_variants=False,
-        is_shipping_required=False,
-    )
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-4",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    variant = ProductVariant.objects.create(product=product, sku="SKU_C")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(0),
-        discounted_price_amount=Decimal(0),
-        cost_price_amount=Decimal(0),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(product_variant=variant, warehouse=warehouse, quantity=1)
-    return product
-
-
-@pytest.fixture
-def product_in_channel_JPY(product, channel_JPY, warehouse_JPY):
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_JPY,
-        is_published=True,
-        discounted_price_amount="1200",
-        currency=channel_JPY.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    variant = product.variants.get()
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_JPY,
-        price_amount=Decimal(1200),
-        discounted_price_amount=Decimal(1200),
-        cost_price_amount=Decimal(300),
-        currency=channel_JPY.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse_JPY, product_variant=variant, quantity=10)
-    return product
-
-
-@pytest.fixture
-def non_shippable_gift_card_product(
-    non_shippable_gift_card_product_type, category, warehouse, channel_USD
-):
-    product_type = non_shippable_gift_card_product_type
-
-    product = Product.objects.create(
-        name="Non shippable gift card",
-        slug="non-shippable-gift-card",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        discounted_price_amount="200.00",
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variant = ProductVariant.objects.create(
-        product=product, sku="785", track_inventory=False
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(250),
-        discounted_price_amount=Decimal(250),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=1)
-
-    return product
-
-
-@pytest.fixture
-def product_with_rich_text_attribute(
-    product_type_with_rich_text_attribute, category, warehouse, channel_USD
-):
-    product_attr = product_type_with_rich_text_attribute.product_attributes.first()
-    product_attr_value = product_attr.values.first()
-
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-11",
-        product_type=product_type_with_rich_text_attribute,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        discounted_price_amount="10.00",
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    associate_attribute_values_to_instance(
-        product, {product_attr.pk: [product_attr_value]}
-    )
-
-    variant_attr = product_type_with_rich_text_attribute.variant_attributes.first()
-    variant_attr_value = variant_attr.values.first()
-
-    variant = ProductVariant.objects.create(product=product, sku="123")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=10)
-
-    associate_attribute_values_to_instance(
-        variant, {variant_attr.pk: [variant_attr_value]}
-    )
-    return [product, variant]
-
-
-@pytest.fixture
-def product_with_collections(
-    product, published_collection, unpublished_collection, collection
-):
-    product.collections.add(*[published_collection, unpublished_collection, collection])
-    return product
-
-
-@pytest.fixture
-def product_available_in_many_channels(product, channel_PLN, channel_USD):
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_PLN,
-        is_published=True,
-    )
-    variant = product.variants.get()
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_PLN,
-        price_amount=Decimal(50),
-        discounted_price_amount=Decimal(50),
-        cost_price_amount=Decimal(1),
-        currency=channel_PLN.currency_code,
-    )
-    return product
-
-
-@pytest.fixture
-def product_with_single_variant(product_type, category, warehouse, channel_USD):
-    product = Product.objects.create(
-        name="Test product with single variant",
-        slug="test-product-with-single-variant",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    variant = ProductVariant.objects.create(product=product, sku="SKU_SINGLE_VARIANT")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(1.99),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(product_variant=variant, warehouse=warehouse, quantity=101)
-    return product
-
-
-@pytest.fixture
-def product_with_two_variants(product_type, category, warehouse, channel_USD):
-    product = Product.objects.create(
-        name="Test product with two variants",
-        slug="test-product-with-two-variant",
-        product_type=product_type,
-        category=category,
-    )
-
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variants = [
-        ProductVariant(
-            product=product,
-            sku=f"Product variant #{i}",
-        )
-        for i in (1, 2)
-    ]
-    ProductVariant.objects.bulk_create(variants)
-    variants_channel_listing = [
-        ProductVariantChannelListing(
-            variant=variant,
-            channel=channel_USD,
-            price_amount=Decimal(10),
-            discounted_price_amount=Decimal(10),
-            cost_price_amount=Decimal(1),
-            currency=channel_USD.currency_code,
-        )
-        for variant in variants
-    ]
-    ProductVariantChannelListing.objects.bulk_create(variants_channel_listing)
-    Stock.objects.bulk_create(
-        [
-            Stock(
-                warehouse=warehouse,
-                product_variant=variant,
-                quantity=10,
-            )
-            for variant in variants
-        ]
-    )
-    product.search_vector = FlatConcatSearchVector(
-        *prepare_product_search_vector_value(product)
-    )
-    product.save(update_fields=["search_vector"])
-
-    return product
-
-
-@pytest.fixture
-def product_with_variant_with_two_attributes(
-    color_attribute, size_attribute, category, warehouse, channel_USD
-):
-    product_type = ProductType.objects.create(
-        name="Type with two variants",
-        slug="two-variants",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.variant_attributes.add(color_attribute)
-    product_type.variant_attributes.add(size_attribute)
-
-    product = Product.objects.create(
-        name="Test product with two variants",
-        slug="test-product-with-two-variant",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variant = ProductVariant.objects.create(product=product, sku="prodVar1")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-
-    associate_attribute_values_to_instance(
-        variant, {color_attribute.pk: [color_attribute.values.first()]}
-    )
-    associate_attribute_values_to_instance(
-        variant, {size_attribute.pk: [size_attribute.values.first()]}
-    )
-
-    return product
-
-
-@pytest.fixture
-def product_with_variant_with_external_media(
-    color_attribute,
-    size_attribute,
-    category,
-    warehouse,
-    channel_USD,
-):
-    product_type = ProductType.objects.create(
-        name="Type with two variants",
-        slug="two-variants",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.variant_attributes.add(color_attribute)
-    product_type.variant_attributes.add(size_attribute)
-
-    product = Product.objects.create(
-        name="Test product with two variants",
-        slug="test-product-with-two-variant",
-        product_type=product_type,
-        category=category,
-    )
-    media_obj = ProductMedia.objects.create(
-        product=product,
-        external_url="https://www.youtube.com/watch?v=di8_dJ3Clyo",
-        alt="video_1",
-        type=ProductMediaTypes.VIDEO,
-        oembed_data="{}",
-    )
-    product.media.add(media_obj)
-
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variant = ProductVariant.objects.create(product=product, sku="prodVar1")
-    variant.media.add(media_obj)
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-
-    associate_attribute_values_to_instance(
-        variant, {color_attribute.pk: [color_attribute.values.first()]}
-    )
-    associate_attribute_values_to_instance(
-        variant, {size_attribute.pk: [size_attribute.values.first()]}
-    )
-
-    return product
-
-
-@pytest.fixture
-def product_with_variant_with_file_attribute(
-    color_attribute, file_attribute, category, warehouse, channel_USD
-):
-    product_type = ProductType.objects.create(
-        name="Type with variant and file attribute",
-        slug="type-with-file-attribute",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=True,
-    )
-    product_type.variant_attributes.add(file_attribute)
-
-    product = Product.objects.create(
-        name="Test product with variant and file attribute",
-        slug="test-product-with-variant-and-file-attribute",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        currency=channel_USD.currency_code,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-
-    variant = ProductVariant.objects.create(
-        product=product,
-        sku="prodVarTest",
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-
-    associate_attribute_values_to_instance(
-        variant, {file_attribute.pk: [file_attribute.values.first()]}
-    )
-
-    return product
-
-
-@pytest.fixture
-def product_with_multiple_values_attributes(product, product_type) -> Product:
-    attribute = Attribute.objects.create(
-        slug="modes",
-        name="Available Modes",
-        input_type=AttributeInputType.MULTISELECT,
-        type=AttributeType.PRODUCT_TYPE,
-    )
-
-    attr_val_1 = AttributeValue.objects.create(
-        attribute=attribute, name="Eco Mode", slug="eco"
-    )
-    attr_val_2 = AttributeValue.objects.create(
-        attribute=attribute, name="Performance Mode", slug="power"
-    )
-
-    product_type.product_attributes.clear()
-    product_type.product_attributes.add(attribute)
-
-    associate_attribute_values_to_instance(
-        product, {attribute.pk: [attr_val_1, attr_val_2]}
-    )
-    return product
-
-
-@pytest.fixture
-def product_with_default_variant(
-    product_type_without_variant, category, warehouse, channel_USD
-):
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-3",
-        product_type=product_type_without_variant,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    variant = ProductVariant.objects.create(
-        product=product, sku="1234", track_inventory=True
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=100)
-
-    product.search_vector = FlatConcatSearchVector(
-        *prepare_product_search_vector_value(product)
-    )
-    product.save(update_fields=["search_vector"])
-
-    return product
-
-
-@pytest.fixture
-def variant_without_inventory_tracking(
-    product_type_without_variant, category, warehouse, channel_USD
-):
-    product = Product.objects.create(
-        name="Test product without inventory tracking",
-        slug="test-product-without-tracking",
-        product_type=product_type_without_variant,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime.now(tz=datetime.UTC),
-    )
-    variant = ProductVariant.objects.create(
-        product=product,
-        sku="tracking123",
-        track_inventory=False,
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(warehouse=warehouse, product_variant=variant, quantity=0)
-    return variant
-
-
-@pytest.fixture
-def variant(product, channel_USD) -> ProductVariant:
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_A", external_reference="SKU_A"
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    return product_variant
-
-
-@pytest.fixture
-def variant_with_translations(variant):
-    variant.translations.create(language_code="pl", name="OldVariant PL")
-    variant.translations.create(language_code="de", name="OldVariant DE")
-    return variant
-
-
-@pytest.fixture
-def variant_with_image(variant, image_list, media_root):
-    media = ProductMedia.objects.create(product=variant.product, image=image_list[0])
-    VariantMedia.objects.create(variant=variant, media=media)
-    return variant
-
-
-@pytest.fixture
-def variant_with_many_stocks(variant, warehouses_with_shipping_zone):
-    warehouses = warehouses_with_shipping_zone
-    Stock.objects.bulk_create(
-        [
-            Stock(warehouse=warehouses[0], product_variant=variant, quantity=4),
-            Stock(warehouse=warehouses[1], product_variant=variant, quantity=3),
-        ]
-    )
-    return variant
-
-
-@pytest.fixture
-def variant_on_promotion(
-    product, channel_USD, promotion_rule, warehouse
-) -> ProductVariant:
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_A", external_reference="SKU_A"
-    )
-    price_amount = Decimal(10)
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=price_amount,
-        discounted_price_amount=price_amount,
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(
-        warehouse=warehouse, product_variant=product_variant, quantity=10
-    )
-
-    promotion_rule.variants.add(product_variant)
-    reward_value = promotion_rule.reward_value
-    discount_amount = price_amount * reward_value / 100
-
-    variant_channel_listing = product_variant.channel_listings.get(channel=channel_USD)
-
-    variant_channel_listing.discounted_price_amount = (
-        variant_channel_listing.price_amount - reward_value
-    )
-    variant_channel_listing.save(update_fields=["discounted_price_amount"])
-
-    variant_channel_listing.variantlistingpromotionrule.create(
-        promotion_rule=promotion_rule,
-        discount_amount=discount_amount,
-        currency=channel_USD.currency_code,
-    )
-
-    return product_variant
-
-
-@pytest.fixture
-def preorder_variant_global_threshold(product, channel_USD):
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_A_P", is_preorder=True, preorder_global_threshold=10
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    return product_variant
-
-
-@pytest.fixture
-def preorder_variant_channel_threshold(product, channel_USD):
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_B_P", is_preorder=True, preorder_global_threshold=None
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-        preorder_quantity_threshold=10,
-    )
-    return product_variant
-
-
-@pytest.fixture
-def preorder_variant_global_and_channel_threshold(product, channel_USD, channel_PLN):
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_C_P", is_preorder=True, preorder_global_threshold=10
-    )
-    ProductVariantChannelListing.objects.bulk_create(
-        [
-            ProductVariantChannelListing(
-                variant=product_variant,
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-                preorder_quantity_threshold=8,
-            ),
-            ProductVariantChannelListing(
-                variant=product_variant,
-                channel=channel_PLN,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                currency=channel_PLN.currency_code,
-                preorder_quantity_threshold=4,
-            ),
-        ]
-    )
-    return product_variant
-
-
-@pytest.fixture
-def preorder_variant_with_end_date(product, channel_USD):
-    product_variant = ProductVariant.objects.create(
-        product=product,
-        sku="SKU_D_P",
-        is_preorder=True,
-        preorder_global_threshold=10,
-        preorder_end_date=timezone.now() + datetime.timedelta(days=10),
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    return product_variant
-
-
-@pytest.fixture
-def variant_with_many_stocks_different_shipping_zones(
-    variant, warehouses_with_different_shipping_zone
-):
-    warehouses = warehouses_with_different_shipping_zone
-    Stock.objects.bulk_create(
-        [
-            Stock(warehouse=warehouses[0], product_variant=variant, quantity=4),
-            Stock(warehouse=warehouses[1], product_variant=variant, quantity=3),
-        ]
-    )
-    return variant
-
-
-@pytest.fixture
-def gift_card_shippable_variant(shippable_gift_card_product, channel_USD, warehouse):
-    product = shippable_gift_card_product
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_CARD_A", track_inventory=False
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(
-        warehouse=warehouse, product_variant=product_variant, quantity=1
-    )
-    return product_variant
-
-
-@pytest.fixture
-def gift_card_non_shippable_variant(
-    non_shippable_gift_card_product, channel_USD, warehouse
-):
-    product = non_shippable_gift_card_product
-    product_variant = ProductVariant.objects.create(
-        product=product, sku="SKU_CARD_B", track_inventory=False
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(
-        warehouse=warehouse, product_variant=product_variant, quantity=1
-    )
-    return product_variant
-
-
-@pytest.fixture
-def product_variant_list(product, channel_USD, channel_PLN):
-    variants = list(
-        ProductVariant.objects.bulk_create(
-            [
-                ProductVariant(product=product, sku="1"),
-                ProductVariant(product=product, sku="2"),
-                ProductVariant(product=product, sku="3"),
-                ProductVariant(product=product, sku="4"),
-            ]
-        )
-    )
-    ProductVariantChannelListing.objects.bulk_create(
-        [
-            ProductVariantChannelListing(
-                variant=variants[0],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                discounted_price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[1],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                discounted_price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[2],
-                channel=channel_PLN,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                discounted_price_amount=Decimal(10),
-                currency=channel_PLN.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[3],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                discounted_price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-            ),
-        ]
-    )
-    return variants
-
-
-@pytest.fixture
-def product_without_shipping(category, warehouse, channel_USD):
-    product_type = ProductType.objects.create(
-        name="Type with no shipping",
-        slug="no-shipping",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=False,
-        is_shipping_required=False,
-    )
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-4",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    variant = ProductVariant.objects.create(product=product, sku="SKU_E")
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(product_variant=variant, warehouse=warehouse, quantity=1)
-    return product
-
-
-@pytest.fixture
-def product_without_category(product):
-    product.category = None
-    product.save()
-    product.channel_listings.all().update(is_published=False)
-    return product
-
-
-@pytest.fixture
-def product_list(
-    product_type, category, warehouse, channel_USD, channel_PLN, default_tax_class
-):
-    product_attr = product_type.product_attributes.first()
-    attr_value = product_attr.values.first()
-
-    products = list(
-        Product.objects.bulk_create(
-            [
-                Product(
-                    name="Test product 1",
-                    slug="test-product-a",
-                    description_plaintext="big blue product",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-                Product(
-                    name="Test product 2",
-                    slug="test-product-b",
-                    description_plaintext="big orange product",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-                Product(
-                    name="Test product 3",
-                    slug="test-product-c",
-                    description_plaintext="small red",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-            ]
-        )
-    )
-    ProductChannelListing.objects.bulk_create(
-        [
-            ProductChannelListing(
-                product=products[0],
-                channel=channel_USD,
-                is_published=True,
-                discounted_price_amount=10,
-                currency=channel_USD.currency_code,
-                visible_in_listings=True,
-                available_for_purchase_at=(
-                    datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC)
-                ),
-            ),
-            ProductChannelListing(
-                product=products[1],
-                channel=channel_USD,
-                is_published=True,
-                discounted_price_amount=20,
-                currency=channel_USD.currency_code,
-                visible_in_listings=True,
-                available_for_purchase_at=(
-                    datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC)
-                ),
-            ),
-            ProductChannelListing(
-                product=products[2],
-                channel=channel_USD,
-                is_published=True,
-                discounted_price_amount=30,
-                currency=channel_USD.currency_code,
-                visible_in_listings=True,
-                available_for_purchase_at=(
-                    datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC)
-                ),
-            ),
-        ]
-    )
-    variants = list(
-        ProductVariant.objects.bulk_create(
-            [
-                ProductVariant(
-                    product=products[0],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-                ProductVariant(
-                    product=products[1],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-                ProductVariant(
-                    product=products[2],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-            ]
-        )
-    )
-    ProductVariantChannelListing.objects.bulk_create(
-        [
-            ProductVariantChannelListing(
-                variant=variants[0],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                discounted_price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[1],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(20),
-                discounted_price_amount=Decimal(20),
-                currency=channel_USD.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[2],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(30),
-                discounted_price_amount=Decimal(30),
-                currency=channel_USD.currency_code,
-            ),
-        ]
-    )
-    stocks = []
-    for variant in variants:
-        stocks.append(Stock(warehouse=warehouse, product_variant=variant, quantity=100))
-    Stock.objects.bulk_create(stocks)
-
-    for product in products:
-        associate_attribute_values_to_instance(product, {product_attr.pk: [attr_value]})
-        product.search_vector = FlatConcatSearchVector(
-            *prepare_product_search_vector_value(product)
-        )
-
-    Product.objects.bulk_update(products, ["search_vector"])
-
-    return products
-
-
-@pytest.fixture
-def product_list_with_variants_many_channel(
-    product_type, category, channel_USD, channel_PLN, default_tax_class
-):
-    products = list(
-        Product.objects.bulk_create(
-            [
-                Product(
-                    name="Test product 1",
-                    slug="test-product-a",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-                Product(
-                    name="Test product 2",
-                    slug="test-product-b",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-                Product(
-                    name="Test product 3",
-                    slug="test-product-c",
-                    category=category,
-                    product_type=product_type,
-                    tax_class=default_tax_class,
-                ),
-            ]
-        )
-    )
-    ProductChannelListing.objects.bulk_create(
-        [
-            # Channel: USD
-            ProductChannelListing(
-                product=products[0],
-                channel=channel_USD,
-                is_published=True,
-                currency=channel_USD.currency_code,
-                visible_in_listings=True,
-            ),
-            # Channel: PLN
-            ProductChannelListing(
-                product=products[1],
-                channel=channel_PLN,
-                is_published=True,
-                currency=channel_PLN.currency_code,
-                visible_in_listings=True,
-            ),
-            ProductChannelListing(
-                product=products[2],
-                channel=channel_PLN,
-                is_published=True,
-                currency=channel_PLN.currency_code,
-                visible_in_listings=True,
-            ),
-        ]
-    )
-    variants = list(
-        ProductVariant.objects.bulk_create(
-            [
-                ProductVariant(
-                    product=products[0],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-                ProductVariant(
-                    product=products[1],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-                ProductVariant(
-                    product=products[2],
-                    sku=str(uuid.uuid4()).replace("-", ""),
-                    track_inventory=True,
-                ),
-            ]
-        )
-    )
-    ProductVariantChannelListing.objects.bulk_create(
-        [
-            # Channel: USD
-            ProductVariantChannelListing(
-                variant=variants[0],
-                channel=channel_USD,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(10),
-                currency=channel_USD.currency_code,
-            ),
-            # Channel: PLN
-            ProductVariantChannelListing(
-                variant=variants[1],
-                channel=channel_PLN,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(20),
-                currency=channel_PLN.currency_code,
-            ),
-            ProductVariantChannelListing(
-                variant=variants[2],
-                channel=channel_PLN,
-                cost_price_amount=Decimal(1),
-                price_amount=Decimal(30),
-                currency=channel_PLN.currency_code,
-            ),
-        ]
-    )
-    return products
-
-
-@pytest.fixture
-def product_list_with_many_channels(product_list, channel_PLN):
-    ProductChannelListing.objects.bulk_create(
-        [
-            ProductChannelListing(
-                product=product_list[0],
-                channel=channel_PLN,
-                is_published=True,
-            ),
-            ProductChannelListing(
-                product=product_list[1],
-                channel=channel_PLN,
-                is_published=True,
-            ),
-            ProductChannelListing(
-                product=product_list[2],
-                channel=channel_PLN,
-                is_published=True,
-            ),
-        ]
-    )
-    return product_list
-
-
-@pytest.fixture
-def product_list_unpublished(product_list, channel_USD):
-    products = Product.objects.filter(pk__in=[product.pk for product in product_list])
-    ProductChannelListing.objects.filter(
-        product__in=products, channel=channel_USD
-    ).update(is_published=False)
-    return products
-
-
-@pytest.fixture
-def product_list_published(product_list, channel_USD):
-    products = Product.objects.filter(pk__in=[product.pk for product in product_list])
-    ProductChannelListing.objects.filter(
-        product__in=products, channel=channel_USD
-    ).update(is_published=True)
-    return products
-
-
-@pytest.fixture
-def product_with_image(product, image, media_root):
-    ProductMedia.objects.create(product=product, image=image)
-    return product
-
-
-@pytest.fixture
-def product_with_image_list(product, image_list, media_root):
-    ProductMedia.objects.create(product=product, image=image_list[0])
-    ProductMedia.objects.create(product=product, image=image_list[1])
-    return product
-
-
-@pytest.fixture
-def product_with_image_list_and_one_null_sort_order(product_with_image_list):
-    """Return a product with mixed sorting order.
-
-    As we allow to have `null` in `sort_order` in database, but our logic
-    covers changing any new `null` values to proper `int` need to execute
-    raw SQL query on database to test behavior of `null` `sort_order`.
-
-    SQL query behavior:
-    Updates one of the product media `sort_order` to `null`.
-    """
-    with connection.cursor() as cursor:
-        cursor.execute(
-            """
-            UPDATE PRODUCT_PRODUCTMEDIA
-            SET SORT_ORDER = NULL
-            WHERE ID IN (
-                SELECT ID FROM PRODUCT_PRODUCTMEDIA
-                WHERE PRODUCT_ID = %s
-                ORDER BY ID
-                LIMIT 1
-            )
-            """,
-            [product_with_image_list.pk],
-        )
-    product_with_image_list.refresh_from_db()
-    return product_with_image_list
-
-
-@pytest.fixture
-def unavailable_product(product_type, category, channel_USD, default_tax_class):
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-5",
-        product_type=product_type,
-        category=category,
-        tax_class=default_tax_class,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=False,
-        visible_in_listings=False,
-    )
-    return product
-
-
-@pytest.fixture
-def unavailable_product_with_variant(
-    product_type, category, warehouse, channel_USD, default_tax_class
-):
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-6",
-        product_type=product_type,
-        category=category,
-        tax_class=default_tax_class,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=False,
-        visible_in_listings=False,
-    )
-
-    variant_attr = product_type.variant_attributes.first()
-    variant_attr_value = variant_attr.values.first()
-
-    variant = ProductVariant.objects.create(
-        product=product,
-        sku="123",
-    )
-    ProductVariantChannelListing.objects.create(
-        variant=variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(product_variant=variant, warehouse=warehouse, quantity=10)
-
-    associate_attribute_values_to_instance(
-        variant, {variant_attr.pk: [variant_attr_value]}
-    )
-    return product
-
-
-@pytest.fixture
-def product_with_images(
-    product_type, category, media_root, channel_USD, default_tax_class
-):
-    product = Product.objects.create(
-        name="Test product",
-        slug="test-product-7",
-        product_type=product_type,
-        category=category,
-        tax_class=default_tax_class,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-    )
-    file_mock_0 = MagicMock(spec=File, name="FileMock0")
-    file_mock_0.name = "image0.jpg"
-    file_mock_1 = MagicMock(spec=File, name="FileMock1")
-    file_mock_1.name = "image1.jpg"
-    product.media.create(image=file_mock_0)
-    product.media.create(image=file_mock_1)
-    return product
 
 
 @pytest.fixture
@@ -5041,147 +3372,6 @@ def shop_permissions(
 
 
 @pytest.fixture
-def collection(db):
-    collection = Collection.objects.create(
-        name="Collection",
-        slug="collection",
-        description=dummy_editorjs("Test description."),
-    )
-    return collection
-
-
-@pytest.fixture
-def published_collection(db, channel_USD):
-    collection = Collection.objects.create(
-        name="Collection USD",
-        slug="collection-usd",
-        description=dummy_editorjs("Test description."),
-    )
-    CollectionChannelListing.objects.create(
-        channel=channel_USD,
-        collection=collection,
-        is_published=True,
-        published_at=timezone.now(),
-    )
-    return collection
-
-
-@pytest.fixture
-def published_collections(db, channel_USD):
-    collections = Collection.objects.bulk_create(
-        [
-            Collection(
-                name="Collection1",
-                slug="coll1",
-            ),
-            Collection(
-                name="Collection2",
-                slug="coll2",
-            ),
-            Collection(
-                name="Collection3",
-                slug="coll3",
-            ),
-        ]
-    )
-    CollectionChannelListing.objects.bulk_create(
-        [
-            CollectionChannelListing(
-                channel=channel_USD,
-                collection=collection,
-                is_published=True,
-                published_at=datetime.datetime(2019, 4, 10, tzinfo=datetime.UTC),
-            )
-            for collection in collections
-        ]
-    )
-
-    return collections
-
-
-@pytest.fixture
-def published_collection_PLN(db, channel_PLN):
-    collection = Collection.objects.create(
-        name="Collection PLN",
-        slug="collection-pln",
-        description=dummy_editorjs("Test description."),
-    )
-    CollectionChannelListing.objects.create(
-        channel=channel_PLN,
-        collection=collection,
-        is_published=True,
-        published_at=timezone.now(),
-    )
-    return collection
-
-
-@pytest.fixture
-def unpublished_collection(db, channel_USD):
-    collection = Collection.objects.create(
-        name="Unpublished Collection",
-        slug="unpublished-collection",
-        description=dummy_editorjs("Test description."),
-    )
-    CollectionChannelListing.objects.create(
-        channel=channel_USD, collection=collection, is_published=False
-    )
-    return collection
-
-
-@pytest.fixture
-def unpublished_collection_PLN(db, channel_PLN):
-    collection = Collection.objects.create(
-        name="Collection",
-        slug="collection",
-        description=dummy_editorjs("Test description."),
-    )
-    CollectionChannelListing.objects.create(
-        channel=channel_PLN, collection=collection, is_published=False
-    )
-    return collection
-
-
-@pytest.fixture
-def collection_with_products(db, published_collection, product_list_published):
-    published_collection.products.set(list(product_list_published))
-    return product_list_published
-
-
-@pytest.fixture
-def collection_with_image(db, image, media_root, channel_USD):
-    collection = Collection.objects.create(
-        name="Collection",
-        slug="collection",
-        description=dummy_editorjs("Test description."),
-        background_image=image,
-    )
-    CollectionChannelListing.objects.create(
-        channel=channel_USD, collection=collection, is_published=False
-    )
-    return collection
-
-
-@pytest.fixture
-def collection_list(db, channel_USD):
-    collections = Collection.objects.bulk_create(
-        [
-            Collection(name="Collection 1", slug="collection-1"),
-            Collection(name="Collection 2", slug="collection-2"),
-            Collection(name="Collection 3", slug="collection-3"),
-        ]
-    )
-    CollectionChannelListing.objects.bulk_create(
-        [
-            CollectionChannelListing(
-                channel=channel_USD, collection=collection, is_published=True
-            )
-            for collection in collections
-        ]
-    )
-    return collections
-
-
-@pytest.fixture
 def translated_attribute(product):
     attribute = product.product_type.product_attributes.first()
     return AttributeTranslation.objects.create(
@@ -5553,60 +3743,6 @@ def transaction_item(order, transaction_item_generator):
     return transaction_item_generator(
         order_id=order.pk,
     )
-
-
-@pytest.fixture
-def digital_content(category, media_root, warehouse, channel_USD) -> DigitalContent:
-    product_type = ProductType.objects.create(
-        name="Digital Type",
-        slug="digital-type",
-        kind=ProductTypeKind.NORMAL,
-        has_variants=True,
-        is_shipping_required=False,
-        is_digital=True,
-    )
-    product = Product.objects.create(
-        name="Test digital product",
-        slug="test-digital-product",
-        product_type=product_type,
-        category=category,
-    )
-    ProductChannelListing.objects.create(
-        product=product,
-        channel=channel_USD,
-        is_published=True,
-        visible_in_listings=True,
-        available_for_purchase_at=datetime.datetime(1999, 1, 1, tzinfo=datetime.UTC),
-    )
-    product_variant = ProductVariant.objects.create(product=product, sku="SKU_554")
-    ProductVariantChannelListing.objects.create(
-        variant=product_variant,
-        channel=channel_USD,
-        price_amount=Decimal(10),
-        discounted_price_amount=Decimal(10),
-        cost_price_amount=Decimal(1),
-        currency=channel_USD.currency_code,
-    )
-    Stock.objects.create(
-        product_variant=product_variant,
-        warehouse=warehouse,
-        quantity=5,
-    )
-
-    assert product_variant.is_digital()
-
-    image_file, image_name = create_image()
-    d_content = DigitalContent.objects.create(
-        content_file=image_file,
-        product_variant=product_variant,
-        use_default_settings=True,
-    )
-    return d_content
-
-
-@pytest.fixture
-def digital_content_url(digital_content, order_line):
-    return DigitalContentUrl.objects.create(content=digital_content, line=order_line)
 
 
 @pytest.fixture
@@ -6314,17 +4450,6 @@ def success_gateway_response():
 
 
 @pytest.fixture
-def product_media_image(product, image, media_root):
-    return ProductMedia.objects.create(
-        product=product,
-        image=image,
-        alt="image",
-        type=ProductMediaTypes.IMAGE,
-        oembed_data="{}",
-    )
-
-
-@pytest.fixture
 def transaction_session_response():
     return {
         "pspReference": "psp-123",
@@ -6827,62 +4952,6 @@ def async_subscription_webhooks_with_root_objects(
             warehouse,
         ],
     }
-
-
-@pytest.fixture
-def lots_of_products_with_variants(product_type, channel_USD):
-    def chunks(iterable, size):
-        it = iter(iterable)
-        chunk = tuple(itertools.islice(it, size))
-        while chunk:
-            yield chunk
-            chunk = tuple(itertools.islice(it, size))
-
-    variants_per_product = 4
-    products_count = 10000
-    slug_generator = (f"test-slug-{i}" for i in range(products_count))
-
-    for batch in chunks(range(products_count), 500):
-        batch_len = len(batch)
-        variants = []
-        product_listings = []
-        products = [
-            Product(
-                name=i,
-                slug=next(slug_generator),
-                product_type_id=product_type.pk,
-            )
-            for i in range(batch_len)
-        ]
-        for product in Product.objects.bulk_create(products):
-            product_listings.append(
-                ProductChannelListing(
-                    channel=channel_USD,
-                    product=product,
-                    visible_in_listings=True,
-                    available_for_purchase_at="2022-01-01",
-                    currency=channel_USD.currency_code,
-                )
-            )
-            for x in range(variants_per_product):
-                variant = ProductVariant(name=x, product_id=product.id)
-                variants.append(variant)
-        ProductVariant.objects.bulk_create(variants)
-        variant_listings = []
-        for variant in variants:
-            price = random.randint(1, 100)
-            variant_listings.append(
-                ProductVariantChannelListing(
-                    variant=variant,
-                    channel=channel_USD,
-                    currency=channel_USD.currency_code,
-                    price_amount=price,
-                    discounted_price_amount=price,
-                )
-            )
-        ProductVariantChannelListing.objects.bulk_create(variant_listings)
-        ProductChannelListing.objects.bulk_create(product_listings)
-    return Product.objects.all()
 
 
 @pytest.fixture
